seek to current time on 'play'
Safari is the only browser that allows you to seek before play is ready,
it seems. All other browsers require you to seek a bit later in the
event timeline. I am not sure if it can be seeked slightly earlier, but
this seems to work nicely across all browsers for now.